Move fake controllers to `stubs`
