+++ tests/framework/console/ControllerTest.php
@@ -18,6 +18,10 @@ use yii\base\Module;
 use yii\console\Application;
 use yii\console\Request;
 use yii\helpers\Console;
+use yiiunit\framework\console\stubs\FakeController;
+use yiiunit\framework\console\stubs\FakeHelpController;
+use yiiunit\framework\console\stubs\FakeHelpControllerWithoutOutput;
+use yiiunit\framework\console\stubs\FakePhp71Controller;
 use yiiunit\TestCase;
 
 /**
@@ -33,9 +37,9 @@ class ControllerTest extends TestCase
         parent::setUp();
         $this->mockApplication();
         Yii::$app->controllerMap = [
-            'fake' => 'yiiunit\framework\console\FakeController',
-            'fake_witout_output' => 'yiiunit\framework\console\FakeHelpControllerWithoutOutput',
-            'help' => 'yiiunit\framework\console\FakeHelpController',
+            'fake' => 'yiiunit\framework\console\stubs\FakeController',
+            'fake_witout_output' => 'yiiunit\framework\console\stubs\FakeHelpControllerWithoutOutput',
+            'help' => 'yiiunit\framework\console\stubs\FakeHelpController',
         ];
     }
