+++ tests/framework/console/stubs/FakeController.php
@@ -6,7 +6,7 @@
  * @license https://www.yiiframework.com/license/
  */
 
-namespace yiiunit\framework\console;
+namespace yiiunit\framework\console\stubs;
 
 use yii\console\Controller;
 use yii\console\Response;
