+++ tests/framework/web/ControllerTest.php
@@ -20,6 +20,10 @@ use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\web\Response;
 use yii\web\ServerErrorHttpException;
+use yiiunit\framework\web\stubs\FakeController;
+use yiiunit\framework\web\stubs\FakePhp7Controller;
+use yiiunit\framework\web\stubs\FakePhp71Controller;
+use yiiunit\framework\web\stubs\FakePhp80Controller;
 use yiiunit\framework\web\stubs\VendorImage;
 use yiiunit\TestCase;
