+++ tests/framework/web/stubs/FakePhp71Controller.php
@@ -6,21 +6,26 @@
  * @license https://www.yiiframework.com/license/
  */
 
-namespace yiiunit\framework\console;
+namespace yiiunit\framework\web\stubs;
 
 use yii\data\DataProviderInterface;
-use yiiunit\framework\console\stubs\DummyService;
-use yii\console\Controller;
-use yii\console\Request;
-use yiiunit\framework\base\Post;
+use yii\web\Controller;
+use yii\web\Request;
+use yiiunit\framework\web\Post;
 
+/**
+ * @author Sam Mousa<sam@mousa.nl>
+ * @since 2.0.36
+ */
 class FakePhp71Controller extends Controller
 {
+    public $enableCsrfValidation = false;
+
     public function actionInjection(
         $before,
         Request $request,
         $between,
-        DummyService $dummyService,
+        VendorImage $vendorImage,
         ?Post $post,
         $after
     ) {
@@ -31,6 +36,10 @@ class FakePhp71Controller extends Controller
     }
 
     public function actionModuleServiceInjection(DataProviderInterface $dataProvider)
+    {
+    }
+
+    public function actionModelBindingInjection(ModelBindingStub $model)
     {
     }
 }
